chore: move dependent services code behind windows build tag
